+++ packages/components/src/ui/Dialog/index.ts
@@ -1,1 +1,2 @@
+export * from "./Dialog";
 export * from "./DodamDialog";
